Working unit tests and finished implementation, one bug found with no basket found but product stock updated.

+++ akka-microservices-proj/Domain/BasketProduct.cs
@@ -9,7 +9,7 @@ namespace akka_microservices_proj.Domain
     {
         public long BasketProductId { get; set; }
         public string Name { get; set; }
-        public decimal Price { get; set; }
+        public int Price { get; set; }
         public int AmountAdded { get; set; }
         public int AmountRemoved { get; set; }
     }
